requirements.txt
python
mysql-connector-python
pytest